Change codebase to the Nspire's version. thanks rwill. Consolidating source code

+++ source/d_sprite.c
@@ -213,14 +213,14 @@ void D_SpriteScanLeftEdge (void)
 	if (lmaxindex == 0)
 		lmaxindex = r_spritedesc.nump;
 
-	vtop = (float)ceil (r_spritedesc.pverts[i].v);
+	vtop = ceilf (r_spritedesc.pverts[i].v);
 
 	do
 	{
 		pvert = &r_spritedesc.pverts[i];
 		pnext = pvert - 1;
 
-		vbottom = (float)ceil (pnext->v);
+		vbottom = ceilf (pnext->v);
 
 		if (vtop < vbottom)
 		{
@@ -252,6 +252,7 @@ void D_SpriteScanLeftEdge (void)
 	} while (i != lmaxindex);
 }
 
+
 /*
 =====================
 D_SpriteScanRightEdge
@@ -274,7 +275,7 @@ void D_SpriteScanRightEdge (void)
 	if (vvert > r_refdef.fvrectbottom_adj)
 		vvert = r_refdef.fvrectbottom_adj;
 
-	vtop = (float)ceil (vvert);
+	vtop = ceilf (vvert);
 
 	do
 	{
@@ -287,7 +288,7 @@ void D_SpriteScanRightEdge (void)
 		if (vnext > r_refdef.fvrectbottom_adj)
 			vnext = r_refdef.fvrectbottom_adj;
 
-		vbottom = (float)ceil (vnext);
+		vbottom = ceilf (vnext);
 
 		if (vtop < vbottom)
 		{
@@ -333,6 +334,7 @@ void D_SpriteScanRightEdge (void)
 	pspan->count = DS_SPAN_LIST_END;	// mark the end of the span list 
 }
 
+
 /*
 =====================
 D_SpriteCalculateGradients
@@ -348,7 +350,7 @@ void D_SpriteCalculateGradients (void)
 	TransformVector (r_spritedesc.vup, p_taxis);
 	VectorInverse (p_taxis);
 
-	distinv = (float)(1.0 / (-DotProduct (modelorg, r_spritedesc.vpn)));
+	distinv = 1.0 / (-DotProduct (modelorg, r_spritedesc.vpn));
 
 	d_sdivzstepu = p_saxis[0] * xscaleinv;
 	d_tdivzstepu = p_taxis[0] * xscaleinv;
@@ -378,6 +380,7 @@ void D_SpriteCalculateGradients (void)
 	bbextentt = (sprite_height << 16) - 1;
 }
 
+
 /*
 =====================
 D_DrawSprite
@@ -394,8 +397,8 @@ void D_DrawSprite (void)
 
 // find the top and bottom vertices, and make sure there's at least one scan to
 // draw
-	ymin = (float)999999.9;
-	ymax = (float)-999999.9;
+	ymin = 999999.9;
+	ymax = -999999.9;
 	pverts = r_spritedesc.pverts;
 
 	for (i=0 ; i<r_spritedesc.nump ; i++)
@@ -415,8 +418,8 @@ void D_DrawSprite (void)
 		pverts++;
 	}
 
-	ymin = (float)ceil (ymin);
-	ymax = (float)ceil (ymax);
+	ymin = ceilf (ymin);
+	ymax = ceilf (ymax);
 
 	if (ymin >= ymax)
 		return;		// doesn't cross any scans at all
@@ -436,3 +439,4 @@ void D_DrawSprite (void)
 	D_SpriteScanRightEdge ();
 	D_SpriteDrawSpans (sprite_spans);
 }
+
