+++ source/wad.c
@@ -73,7 +73,6 @@ void W_LoadWadFile (char *filename)
 	int				infotableofs;
 	
 	wad_base = COM_LoadHunkFile (filename);
-
 	if (!wad_base)
 		Sys_Error ("W_LoadWadFile: couldn't load %s", filename);
 
@@ -84,12 +83,12 @@ void W_LoadWadFile (char *filename)
 	|| header->identification[2] != 'D'
 	|| header->identification[3] != '2')
 		Sys_Error ("Wad file %s doesn't have WAD2 id\n",filename);
-	
+		
 	wad_numlumps = LittleLong(header->numlumps);
 	infotableofs = LittleLong(header->infotableofs);
 	wad_lumps = (lumpinfo_t *)(wad_base + infotableofs);
 	
-	for (i=0, lump_p = wad_lumps ; i<(unsigned int)wad_numlumps ; i++,lump_p++)
+	for (i=0, lump_p = wad_lumps ; i<wad_numlumps ; i++,lump_p++)
 	{
 		lump_p->filepos = LittleLong(lump_p->filepos);
 		lump_p->size = LittleLong(lump_p->size);
